final project

+++ frontend/src/components/WorkoutDetails.js
@@ -1,6 +1,7 @@
 import { useState } from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 import EditWorkoutForm from './EditWorkoutForm';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 
 //date fns
@@ -10,10 +11,17 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 const WorkoutDetails = ({ workout, setWorkouts }) => {
   const { dispatch } = useWorkoutsContext();
   const [isEditing, setIsEditing] = useState(false);
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if(!user){
+      return;
+    }
     const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json();
